Refactors and enhances actor registration
