+++ src/Servus.Akka/DependencyInjection/IActorRef.cs
@@ -1,6 +1,6 @@
 ﻿using Akka.Actor;
 
-namespace Servus.Akka;
+namespace Servus.Akka.DependencyInjection;
 
 public interface IActorRef<TActor> : IActorRef where TActor : ActorBase
 {
