+++ src/Servus.Akka/Startup/AkkaStartupContainer.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.Extensions.Hosting;
+using Servus.Akka.DependencyInjection;
+using Servus.Core.Application.Startup;
+
+namespace Servus.Akka.Startup;
+
+public class AkkaStartupContainer : IHostBuilderSetupContainer
+{
+    public void ConfigureHostBuilder(IHostBuilder builder)
+    {
+        builder.UseServiceProviderFactory(new ActorRefProviderFactory());
+    }
+}
